+++ logs_sdk/lib/opentelemetry/sdk/logs.rb
@@ -5,6 +5,10 @@
 # SPDX-License-Identifier: Apache-2.0
 
 require_relative 'logs/version'
+require_relative 'logs/logger'
+require_relative 'logs/logger_provider'
+require_relative 'logs/log_record_processor'
+require_relative 'logs/export'
 
 module OpenTelemetry
   module SDK